Protegendo as rotas conforme perfil de usuário

+++ src/app/pages/acrescimo/acrescimo.component.ts
@@ -1,3 +1,4 @@
+import { SessaoService } from 'src/app/services/sessao.service';
 import { ModalAcrescimoComponent } from './modal-acrescimo/modal-acrescimo.component';
 import { MatDialog } from '@angular/material/dialog';
 import { AcrescimoService } from './../../services/acrescimo.service';
@@ -14,12 +15,20 @@ export class AcrescimoComponent implements OnInit {
   fonteAcrescimos: IAcrescimo[] = [];
   displayedColumns = ['name', 'price', 'createdAt', 'status', 'actions'];
 
-  constructor(private acrescimoService: AcrescimoService, private dialog: MatDialog) { }
+  constructor(
+    private acrescimoService: AcrescimoService,
+    private dialog: MatDialog,
+    private sessaoService: SessaoService,
+  ) { }
 
   ngOnInit(): void {
     this.acrescimoService.read().subscribe(acrescimos => {
       this.fonteAcrescimos = acrescimos;
     });
+  }
+
+  notPermission(): boolean {
+    return !this.sessaoService.checkPermission();
   }
 
   dialogCadastrar(): void {
@@ -44,45 +53,51 @@ export class AcrescimoComponent implements OnInit {
 
   dialogEditar(event: MouseEvent, acrescimo: IAcrescimo): void {
     event.stopPropagation();
-    const dialogRef = this.dialog.open(ModalAcrescimoComponent, {
-      width: '50%',
-      data: {
-        title: 'Editar acréscimo',
-        id: acrescimo.id,
-        name: acrescimo.name,
-        price: acrescimo.price,
-      }
-    });
 
-    dialogRef.afterClosed().subscribe((result: IAcrescimo) => {
-      if (result) {
-        const acrescimo: IAcrescimo = {
-          name: result.name,
-          price: result.price,
+    if (!this.notPermission()) {
+      const dialogRef = this.dialog.open(ModalAcrescimoComponent, {
+        width: '50%',
+        data: {
+          title: 'Editar acréscimo',
+          id: acrescimo.id,
+          name: acrescimo.name,
+          price: acrescimo.price,
         }
-        this.acrescimoService.update(acrescimo, result.id).subscribe(() => {
-          this.acrescimoService.showMessage('Acréscimo alterado com sucesso!');
-          this.ngOnInit();
-        });
-      }
-    });
+      });
+
+      dialogRef.afterClosed().subscribe((result: IAcrescimo) => {
+        if (result) {
+          const acrescimo: IAcrescimo = {
+            name: result.name,
+            price: result.price,
+          }
+          this.acrescimoService.update(acrescimo, result.id).subscribe(() => {
+            this.acrescimoService.showMessage('Acréscimo alterado com sucesso!');
+            this.ngOnInit();
+          });
+        }
+      });
+    }
   }
 
   dialogExcluir(event: MouseEvent, acrescimo: IAcrescimo): void {
     event.stopPropagation();
-    const dialogRef = this.dialog.open(ModalAcrescimoComponent, {
-      width: '50%',
-      data: { title: 'Excluir acréscimo', id: acrescimo.id, name: acrescimo.name }
-    });
 
-    dialogRef.afterClosed().subscribe((result: IAcrescimo) => {
-      if (result) {
-        this.acrescimoService.delete(result.id).subscribe(() => {
-          this.acrescimoService.showMessage('Acréscimo excluído com sucesso!');
-          this.ngOnInit();
-        });
-      }
-    });
+    if (!this.notPermission()) {
+      const dialogRef = this.dialog.open(ModalAcrescimoComponent, {
+        width: '50%',
+        data: { title: 'Excluir acréscimo', id: acrescimo.id, name: acrescimo.name }
+      });
+
+      dialogRef.afterClosed().subscribe((result: IAcrescimo) => {
+        if (result) {
+          this.acrescimoService.delete(result.id).subscribe(() => {
+            this.acrescimoService.showMessage('Acréscimo excluído com sucesso!');
+            this.ngOnInit();
+          });
+        }
+      });
+    }
   }
 
   ativarDesativar(acrescimo: IAcrescimo): void {
